requirements.txt
blinker==1.5
click==8.1.3
Flask==2.2.2
Flask-Mail==0.9.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
python-http-client==3.3.7
sendgrid==6.9.7
starkbank-ecdsa==2.0.3
Werkzeug==2.2.2
